+++ stockfish-core/src/accelerate/computed.rs
@@ -18,17 +18,17 @@ pub const fn square_distance(s1: Square, s2: Square) -> u8 {
 /// Returns a [`Bitboard`] containing all the [`Square`]s on the same file,
 /// rank, or diagonal as both `s1` and `s2`. Includes `s1` and `s2`.
 pub const fn line(s1: Square, s2: Square) -> Bitboard {
-    if pseudo_attacks(Piece::Bishop, s1).contains(s2) {
+    if pseudo_attacks(Token::Bishop, s1).contains(s2) {
         return (s1 | s2) | (
-            pseudo_attacks(Piece::Bishop, s1) &
-            pseudo_attacks(Piece::Bishop, s2)
+            pseudo_attacks(Token::Bishop, s1) &
+            pseudo_attacks(Token::Bishop, s2)
         );
     }
 
-    if pseudo_attacks(Piece::Rook, s1).contains(s2) {
+    if pseudo_attacks(Token::Rook, s1).contains(s2) {
         return (s1 | s2) | (
-            pseudo_attacks(Piece::Rook, s1) &
-            pseudo_attacks(Piece::Rook, s2)
+            pseudo_attacks(Token::Rook, s1) &
+            pseudo_attacks(Token::Rook, s2)
         );
     }
 
@@ -40,53 +40,53 @@ pub const fn line(s1: Square, s2: Square) -> Bitboard {
 /// the same rank, file, or diagonal, returns `s2`.
 ///
 /// This can allow us to generate non-king evasion moves faster: a defending
-/// piece must either interpose itself to cover the check or capture the
-/// checking piece.
+/// token must either interpose itself to cover the check or capture the
+/// checking token.
 pub const fn between(s1: Square, s2: Square) -> Bitboard {
-    if pseudo_attacks(Piece::Bishop, s1).contains(s2) {
+    if pseudo_attacks(Token::Bishop, s1).contains(s2) {
         return
-            sliding_attacks(Piece::Bishop, s1, s2.into()) &
-            sliding_attacks(Piece::Bishop, s2, s1.into()) |
+            sliding_attacks(Token::Bishop, s1, s2.into()) &
+            sliding_attacks(Token::Bishop, s2, s1.into()) |
             s2;
     }
 
-    if pseudo_attacks(Piece::Rook, s1).contains(s2) {
+    if pseudo_attacks(Token::Rook, s1).contains(s2) {
         return
-            sliding_attacks(Piece::Rook, s1, s2.into()) &
-            sliding_attacks(Piece::Rook, s2, s1.into()) |
+            sliding_attacks(Token::Rook, s1, s2.into()) &
+            sliding_attacks(Token::Rook, s2, s1.into()) |
             s2;
     }
 
     s2.into()
 }
 
-pub const fn attacks(color: Color, piece: Piece, square: Square, occupied: Bitboard) -> Bitboard {
+pub const fn attacks(color: Color, token: Token, square: Square, occupied: Bitboard) -> Bitboard {
     debug_assert!((occupied & square).is_empty(),
-        "occupancy bitboard must not contain the attacking piece");
+        "occupancy bitboard must not contain the attacking token");
 
-    match piece {
-        Piece::Pawn                 => pawn_attacks(color, square),
-        Piece::Knight | Piece::King => pseudo_attacks(piece, square),
-        _                           => sliding_attacks(piece, square, occupied),
+    match token {
+        Token::Pawn                 => pawn_attacks(color, square),
+        Token::Knight | Token::King => pseudo_attacks(token, square),
+        _                           => sliding_attacks(token, square, occupied),
     }
 }
 
-pub const fn pseudo_attacks(piece: Piece, square: Square) -> Bitboard {
+pub const fn pseudo_attacks(token: Token, square: Square) -> Bitboard {
     // pawns require a color to know which direction they attack in
-    debug_assert!(piece != Piece::Pawn,
+    debug_assert!(token != Token::Pawn,
         "pawns do not have pseudo-attacks defined on them");
 
-    // punt to `sliding_attacks` on an empty board for pieces which slide along
+    // punt to `sliding_attacks` on an empty board for tokens which slide along
     // the board (bishop, rook, queen)
-    if piece.is_sliding() {
-        return sliding_attacks(piece, square, Bitboard::EMPTY);
+    if token.is_sliding() {
+        return sliding_attacks(token, square, Bitboard::EMPTY);
     }
 
-    // if the piece doesn't slide, (knight or king), OR together any single
+    // if the token doesn't slide, (knight or king), OR together any single
     // movements that land on a valid square
     let mut i     = 0;
     let mut bb    = Bitboard::EMPTY;
-    let     steps = Piece::STEPS[piece];
+    let     steps = Token::STEPS[token];
 
     while i < steps.len() {
         if let Some(s) = square + steps[i] {
@@ -108,12 +108,12 @@ pub const fn pawn_attacks(color: Color, square: Square) -> Bitboard {
     }
 }
 
-pub const fn sliding_attacks(piece: Piece, square: Square, occupied: Bitboard) -> Bitboard {
-    debug_assert!(piece.is_sliding(),
-        "piece is not capable of sliding attacks");
+pub const fn sliding_attacks(token: Token, square: Square, occupied: Bitboard) -> Bitboard {
+    debug_assert!(token.is_sliding(),
+        "token is not capable of sliding attacks");
 
     let mut attacks    = Bitboard::EMPTY;
-    let     directions = Piece::STEPS[piece];
+    let     directions = Token::STEPS[token];
 
     let mut i = 0;
 
@@ -159,7 +159,7 @@ mod tests {
     #[test]
     fn line_disjoint() {
         for s1 in Square::iter() {
-            let moves = pseudo_attacks(Piece::Queen, s1);
+            let moves = pseudo_attacks(Token::Queen, s1);
 
             for s2 in Square::iter().filter(|s| !moves.contains(*s)) {
                 assert_eq!(Bitboard::EMPTY, line(s1, s2));
@@ -170,7 +170,7 @@ mod tests {
     #[test]
     fn line_bishop_moves() {
         for s1 in Square::iter() {
-            let moves = pseudo_attacks(Piece::Bishop, s1);
+            let moves = pseudo_attacks(Token::Bishop, s1);
 
             for s2 in Square::iter().filter(|s| moves.contains(*s)) {
                 assert!(line(s1, s2).count() >  1);
@@ -183,7 +183,7 @@ mod tests {
     #[test]
     fn line_rook_moves() {
         for s1 in Square::iter() {
-            let moves = pseudo_attacks(Piece::Rook, s1);
+            let moves = pseudo_attacks(Token::Rook, s1);
 
             for s2 in Square::iter().filter(|s| moves.contains(*s)) {
                 assert_eq!(8, line(s1, s2).count());
@@ -201,7 +201,7 @@ mod tests {
     #[test]
     fn between_disjoint() {
         for s1 in Square::iter() {
-            let moves = pseudo_attacks(Piece::Queen, s1);
+            let moves = pseudo_attacks(Token::Queen, s1);
 
             for s2 in Square::iter().filter(|s| !moves.contains(*s)) {
                 assert_eq!(Bitboard::from(s2), between(s1, s2));
@@ -211,9 +211,9 @@ mod tests {
 
     #[test]
     fn between_overlapping() {
-        for piece in [Piece::Bishop, Piece::Rook] {
+        for token in [Token::Bishop, Token::Rook] {
             for s1 in Square::iter() {
-                let moves = pseudo_attacks(piece, s1);
+                let moves = pseudo_attacks(token, s1);
 
                 for s2 in Square::iter().filter(|s| moves.contains(*s)) {
                     assert_eq!(s1.distance(s2) as usize, between(s1, s2).count());
@@ -226,22 +226,22 @@ mod tests {
     fn attacks_pawn() {
         assert_eq!(
             Bitboard::EMPTY,
-            attacks(Color::White, Piece::Pawn, Square::D8, Bitboard::EMPTY),
+            attacks(Color::White, Token::Pawn, Square::D8, Bitboard::EMPTY),
         );
 
         assert_eq!(
             Bitboard::from(Square::B3),
-            attacks(Color::White, Piece::Pawn, Square::A2, Bitboard::EMPTY),
+            attacks(Color::White, Token::Pawn, Square::A2, Bitboard::EMPTY),
         );
 
         assert_eq!(
             Square::D1 | Square::F1,
-            attacks(Color::Black, Piece::Pawn, Square::E2, Bitboard::EMPTY),
+            attacks(Color::Black, Token::Pawn, Square::E2, Bitboard::EMPTY),
         );
 
         assert_eq!(
             Square::G4 | Square::E4,
-            attacks(Color::White, Piece::Pawn, Square::F3, Bitboard::EMPTY | Square::E4 | Square::G4),
+            attacks(Color::White, Token::Pawn, Square::F3, Bitboard::EMPTY | Square::E4 | Square::G4),
         );
     }
 
@@ -250,12 +250,12 @@ mod tests {
         assert_eq!(
             Square::D2 | Square::F2 | Square::C3 | Square::G3 |
             Square::C5 | Square::G5 | Square::D6 | Square::F6,
-            attacks(Color::White, Piece::Knight, Square::E4, Bitboard::ALL ^ Square::E4),
+            attacks(Color::White, Token::Knight, Square::E4, Bitboard::ALL ^ Square::E4),
         );
 
         assert_eq!(
             Square::B3 | Square::C2,
-            attacks(Color::Black, Piece::Knight, Square::A1, Bitboard::EMPTY)
+            attacks(Color::Black, Token::Knight, Square::A1, Bitboard::EMPTY)
         );
     }
 
@@ -263,14 +263,14 @@ mod tests {
     fn attacks_bishop() {
         assert_eq!(
             Square::D3 | Square::F3 | Square::D5 | Square::F5,
-            attacks(Color::White, Piece::Bishop, Square::E4, Bitboard::ALL ^ Square::E4),
+            attacks(Color::White, Token::Bishop, Square::E4, Bitboard::ALL ^ Square::E4),
         );
 
         assert_eq!(
             Square::A1 | Square::C1 | Square::A3 | Square::C3 |
             Square::D4 | Square::E5 | Square::F6 | Square::G7 |
             Square::H8,
-            attacks(Color::Black, Piece::Bishop, Square::B2, Bitboard::EMPTY),
+            attacks(Color::Black, Token::Bishop, Square::B2, Bitboard::EMPTY),
         );
     }
 
@@ -281,45 +281,45 @@ mod tests {
             Square::F5 | Square::F6 |              Square::F8 |
             Square::A7 | Square::B7 | Square::C7 | Square::D7 |
             Square::E7 |              Square::G7 | Square::H7,
-            attacks(Color::White, Piece::Rook, Square::F7, Bitboard::EMPTY),
+            attacks(Color::White, Token::Rook, Square::F7, Bitboard::EMPTY),
         );
 
         assert_eq!(
             Square::F3 | Square::F4 | Square::F5 | Square::F6 |
             Square::D7 | Square::E7 | Square::G7 | Square::H7 |
             Square::F8,
-            attacks(Color::White, Piece::Rook, Square::F7, Bitboard::FILE_D | Bitboard::RANK_3)
+            attacks(Color::White, Token::Rook, Square::F7, Bitboard::FILE_D | Bitboard::RANK_3)
         );
     }
 
     #[test]
-    #[should_panic(expected = "must not contain the attacking piece")]
+    #[should_panic(expected = "must not contain the attacking token")]
     fn attacks_includes_origin_square() {
-        let _ = attacks(Color::White, Piece::King, Square::C7, Square::C7.into());
+        let _ = attacks(Color::White, Token::King, Square::C7, Square::C7.into());
     }
 
     #[test]
     fn pseudo_attacks_with_sliding_piece() {
         assert_eq!(
-            sliding_attacks(Piece::Bishop, Square::A1, Bitboard::EMPTY),
-            pseudo_attacks(Piece::Bishop, Square::A1),
+            sliding_attacks(Token::Bishop, Square::A1, Bitboard::EMPTY),
+            pseudo_attacks(Token::Bishop, Square::A1),
         );
 
         assert_eq!(
-            sliding_attacks(Piece::Rook, Square::H8, Bitboard::EMPTY),
-            pseudo_attacks(Piece::Rook, Square::H8),
+            sliding_attacks(Token::Rook, Square::H8, Bitboard::EMPTY),
+            pseudo_attacks(Token::Rook, Square::H8),
         );
     }
 
     #[test]
     #[should_panic(expected = "pawns do not have pseudo-attacks defined")]
     fn pseudo_attacks_with_pawn() {
-        let _ = pseudo_attacks(Piece::Pawn, Square::D1);
+        let _ = pseudo_attacks(Token::Pawn, Square::D1);
     }
 
     #[test]
     #[should_panic(expected = "not capable of sliding attacks")]
     fn sliding_attacks_must_slide() {
-        let _ = sliding_attacks(Piece::Knight, Square::D4, Bitboard::EMPTY);
+        let _ = sliding_attacks(Token::Knight, Square::D4, Bitboard::EMPTY);
     }
 }