+++ stockfish-game/src/position.rs
@@ -22,11 +22,11 @@ pub struct Position {
     // internal bitboards
     bb_all:      Bitboard,
     bb_by_color: [Bitboard; Color::COUNT],
-    bb_by_piece: [Bitboard; Piece::COUNT],
+    bb_by_piece: [Bitboard; Token::COUNT],
 
     // internal metrics
     count_by_color: [u8; Color::COUNT],
-    count_by_token: [u8; Token::COUNT],
+    count_by_token: [u8; Piece::COUNT],
 
     // TODO: stuff from the StateInfo stockfish struct that eventually doesn't
     // go here
@@ -48,9 +48,9 @@ impl Position {
 
             bb_all:      Bitboard::EMPTY,
             bb_by_color: [Bitboard::EMPTY; Color::COUNT],
-            bb_by_piece: [Bitboard::EMPTY; Piece::COUNT],
+            bb_by_piece: [Bitboard::EMPTY; Token::COUNT],
 
-            count_by_token: [0; Token::COUNT],
+            count_by_token: [0; Piece::COUNT],
             count_by_color: [0; Color::COUNT],
 
             castling_rights: CastlingRights::NONE,
@@ -59,35 +59,35 @@ impl Position {
     }
 
     #[inline]
-    pub fn emplace(&mut self, token: Token, square: Square) {
-        self.board[square] = Some(token);
+    pub fn emplace(&mut self, piece: Piece, square: Square) {
+        self.board[square] = Some(piece);
 
         self.bb_all                     |= square;
-        self.bb_by_color[token.color()] |= square;
-        self.bb_by_piece[token.piece()] |= square;
+        self.bb_by_color[piece.color()] |= square;
+        self.bb_by_piece[piece.token()] |= square;
 
-        self.count_by_color[token.color()] += 1;
-        self.count_by_token[token]         += 1;
+        self.count_by_color[piece.color()] += 1;
+        self.count_by_token[piece]         += 1;
 
         // TODO: piece-square tables
         // psq += PSQT::psq[pc][s];
     }
 
     #[inline]
-    pub fn remove(&mut self, square: Square) -> Option<Token> {
-        let token = self.board[square].take()?;
+    pub fn remove(&mut self, square: Square) -> Option<Piece> {
+        let piece = self.board[square].take()?;
 
         self.bb_all                     ^= square;
-        self.bb_by_piece[token.piece()] ^= square;
-        self.bb_by_color[token.color()] ^= square;
+        self.bb_by_piece[piece.token()] ^= square;
+        self.bb_by_color[piece.color()] ^= square;
 
-        self.count_by_color[token.color()] -= 1;
-        self.count_by_token[token]         -= 1;
+        self.count_by_color[piece.color()] -= 1;
+        self.count_by_token[piece]         -= 1;
 
         // TODO: piece-square tables
         // psq -= PSQT::psq[pc][s];
 
-        Some(token)
+        Some(piece)
     }
 
     #[inline]
@@ -101,18 +101,18 @@ impl Position {
     }
 
     #[inline]
-    pub fn bitboard_for_piece(&self, piece: Piece) -> Bitboard {
-        self.bb_by_piece[piece]
+    pub fn bitboard_for_piece(&self, token: Token) -> Bitboard {
+        self.bb_by_piece[token]
     }
 
     #[inline]
-    pub fn bitboard_for_token(&self, token: Token) -> Bitboard {
-        self.bb_by_color[token.color()] & self.bb_by_piece[token.piece()]
+    pub fn bitboard_for_token(&self, piece: Piece) -> Bitboard {
+        self.bb_by_color[piece.color()] & self.bb_by_piece[piece.token()]
     }
 }
 
 impl Index<Square> for Position {
-    type Output = Option<Token>;
+    type Output = Option<Piece>;
 
     fn index(&self, index: Square) -> &Self::Output {
         self.board.index(index)